requirements.txt
Flask==2.0.2
numpy==1.21.2
librosa==0.8.1
obspy==1.2.2
scikit-learn==0.24.2
joblib==1.0.1
python-dotenv==0.19.0
pymongo==3.11.4
gridfs==3.11.4
whitenoise==5.3.0
pandas==1.3.3